% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_cluster.R
\name{quick_cluster}
\alias{quick_cluster}
\title{Quickly cluster \linkS4class{SingleCellExperiment} object}
\usage{
quick_cluster(sce, method = c("walktrap", "leiden"), label = NULL)
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} object to be clustered.}

\item{method}{A character vector specifying either \code{walktrap} or \code{leiden} graph clustering algorithms.}

\item{label}{A character vector specifying the column in \code{colData} the cluster results should be stored in.
Defaults to \code{method}.}
}
\value{
Clustered \linkS4class{SingleCellExperiment} with clusters saved in \code{label} column of \code{colData}.
}
\description{
Quickly cluster \linkS4class{SingleCellExperiment} object
}
\examples{
sce <- scater::mockSCE() |> scater::logNormCounts() |> scater::runPCA()
sce <- quick_cluster(sce, method = "walktrap", label = "cluster")
str(sce$cluster)
}
