% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_filter.R
\name{quick_filter}
\alias{quick_filter}
\title{Quickly filter cells based on QC metrics}
\usage{
quick_filter(
  sce,
  features = NULL,
  discard = TRUE,
  nmads = 3,
  min_counts = NULL,
  max_counts = NULL
)
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} to be filtered.}

\item{features}{An optional list of features in colData to be filtered on.}

\item{discard}{Logical scalar indicating whether cells not passing the filters should be discarded.}

\item{nmads}{Numeric scalar indicating the of median absolute deviations to filter on.}

\item{min_counts}{Numeric scalar indicating the lower threshold for UMI number.}

\item{max_counts}{Numeric scalar indicating the upper threshold for UMI number.}
}
\value{
A filtered \linkS4class{SingleCellExperiment} (if \code{discard = TRUE}) or a \linkS4class{SingleCellExperiment}
with logical \code{discard} column added to its \code{colData}, indicating which cells did not pass the filters.
}
\description{
Quickly filter cells based on QC metrics
}
\examples{
sce <- scater::mockSCE()
sce.qc <- quick_qc(sce)
sce.filt <- quick_filter(sce.qc)
}
