% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_mnn.R
\name{quick_mnn}
\alias{quick_mnn}
\title{MNN Integrate \linkS4class{SingleCellExperiment} Objects}
\usage{
quick_mnn(...)
}
\arguments{
\item{...}{\linkS4class{SingleCellExperiment} classes to be MNN-integrated.}
}
\value{
\linkS4class{SingleCellExperiment} object with \code{corrected} object in \code{reducedDims}.
}
\description{
MNN Integrate \linkS4class{SingleCellExperiment} Objects
}
\examples{
sce1 <- scater::mockSCE() |> scater::logNormCounts()
sce2 <- scater::mockSCE() |> scater::logNormCounts()
sce.combined <- quick_mnn(sce1, sce2)
}
