% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_qc.R
\name{quick_qc}
\alias{quick_qc}
\title{Perform quick one-step QC}
\usage{
quick_qc(
  sce,
  mito = TRUE,
  ribo = TRUE,
  hemo = TRUE,
  doublets = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} for QC metrics to be calculated.}

\item{mito}{Logical scalar indicating if percentage of mitochondrial reads should be added to colData.}

\item{ribo}{Logical scalar indicating if percentage of ribosomal reads should be added to colData.}

\item{hemo}{Logical scalar indicating if percentage of hemoglobin reads should be added to colData.}

\item{doublets}{Logical scalar indicating if per-cell droplet calculations should be performed.}

\item{verbose}{Logical scalar indicating if outputs should be verbose.}
}
\value{
A \linkS4class{SingleCellExperiment} with QC metrics added to colData.
}
\description{
Perform quick one-step QC
}
\examples{
sce <- scuttle::mockSCE()
sce.qc <- quick_qc(sce, verbose = TRUE)

}
