% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_droplet_processing.R
\name{run_droplet_processing}
\alias{run_droplet_processing}
\title{Load and Perform Droplet Processing on Data From 10x Genomics Dataset}
\usage{
run_droplet_processing(
  fpath,
  out_save = FALSE,
  out_dir = "scrnaseq/",
  plot_save = FALSE,
  force = FALSE,
  verbose = FALSE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{fpath}{A character vector indicating the directory containing the \code{filtered_feature_bc_matrix/} subdirectory.}

\item{out_save}{Logical scalar indicating whether output \linkS4class{SingleCellExperiment} should be written to disk.}

\item{out_dir}{A character vector indicating the directory to save outputs in.}

\item{plot_save}{Logical scalar indicating whether barcode rank plot should be saved to disk.}

\item{force}{Logical scalar indicating that, if the \linkS4class{SingleCellExperiment} output is already found, it should be overwritten.}

\item{verbose}{Logical scalar indicating whether function should intermittently print status.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object passed to relevant Bioconductor functions.}
}
\value{
A \linkS4class{SingleCellExperiment} object containing count data for each gene (row) and cell (column), with empty droplets removed.
}
\description{
Creates a \linkS4class{SingleCellExperiment} from cellranger output directories, and performs simple barcode calling using \code{\link{emptyDrops}}.
}
\examples{
run_droplet_processing(fpath = "/Users/griffen/Documents/thesis-code/scrnaseq/outs/tw2_uninjured/")
}
