% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_soupx.R
\name{run_soupx}
\alias{run_soupx}
\title{Run SoupX decontamination}
\usage{
run_soupx(
  fpath,
  out_save = FALSE,
  out_dir = "scrnaseq/outs_soupx/",
  plot_save = FALSE,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{fpath}{A character vector indicating cellranger \code{outs} directory, which itself which must contain \code{raw_feature_bc_matrix}, \code{filtered_feature_bc_matrix}, and \code{analysis} subdirectories.}

\item{out_save}{Logical scalar indicating whether SoupX-corrected matrix should be written to disk.}

\item{out_dir}{A character vector indicating the directory to save outputs in.}

\item{plot_save}{Logical scalar indicating whether SoupX plot should be saved to disk.}

\item{force}{Logical scalar indicating that, if the SoupX output is already found, it should be overwritten.}

\item{verbose}{Logical scalar indicating whether function should intermittently print status.}
}
\value{
\linkS4class{sparseMatrix} corresponding to SoupX-corrected counts.
}
\description{
Wraps the normal SoupX workflow, with default parameters, into a single
function for convenience, with options as to whether the plots and outputs should be saved to disk.
}
\examples{
fpath <- "/Users/griffen/Documents/thesis-code/scrnaseq/outs/tw2_uninjured/"
soup <- wakelin::run_soupx(fpath = fpath, out_save = FALSE, plot_save = FALSE)

}
