% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_soupx.R
\name{run_soupx}
\alias{run_soupx}
\title{Run SoupX decontamination}
\usage{
run_soupx(
  fpath,
  out_save = FALSE,
  out_dir = "scrnaseq/outs_soupx/",
  plot_save = FALSE,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{fpath}{Path to cellranger \code{outs} directory, which needs to contain \code{raw_feature_bc_matrix}, \code{filtered_feature_bc_matrix}, and \code{analysis} subdirectories.}

\item{out_save}{Save SoupX output matrix?}

\item{out_dir}{Path to output directory where results (and possibly plots) will be saved.}

\item{plot_save}{Save SoupX plot?}

\item{force}{If SoupX output matrix already found, should it be overwritten?}

\item{verbose}{Be verbose?}
}
\value{
SoupX-corrected gene expression matrix.
}
\description{
\code{run_soupx()} wraps the normal SoupX workflow, with default parameters, into a single
function for convenience, with options as to whether the plots and outputs should be saved to disk.
}
\examples{
fpath <- "/Users/griffen/Documents/thesis-code/scrnaseq/outs/tw2_uninjured/"
soup <- wakelin::run_soupx(fpath = fpath, out_save = FALSE, plot_save = FALSE)
}
